refacgor: refactor user model, authentication system, and error handling
- Refined users table schema, renamed fields for consistency (full_name, password_hash)

- Updated SQLx migration and .sqlx directory handling

- Updated UserRepository to return domain model with proper error translation

- Introduced AuthError and UserPersistenceError enums using derive_more for structured error management

- Refactored AuthService to use async registration and login logic, returning error enums

- Implemented login/register endpoints with proper DTOs and API error responses

- Cleaned up unused code and improved code structure for better maintainability

- Adjusted middleware and routes to match new auth logic and error system

- Updated Justfile with SQLx prepare recipe

- Updated Cargo.toml dependencies for error handling and DB features

+++ src/interfaces/repositories.rs
@@ -1,13 +1,15 @@
+use async_trait::async_trait;
+
 use crate::domain::User;
 
-#[async_trait::async_trait]
+#[async_trait]
 pub trait UserRepository: Send + Sync {
     async fn create_user(
         &self,
         email: &str,
         password_hash: &str,
         full_name: &str,
-    ) -> Result<u32, String>;
+    ) -> Result<u64, String>;
 
     async fn get_user_by_email(&self, email: &str) -> Result<Option<User>, String>;
